+++ src/Sitecore.Pathfinder.Core/Languages/BinFiles/BinFileCompiler.cs
@@ -30,7 +30,7 @@ namespace Sitecore.Pathfinder.Languages.BinFiles
 
             try
             {
-                var assembly = Assembly.ReflectionOnlyLoadFrom(binFile.Snapshots.First().SourceFile.AbsoluteFileName);
+                var assembly = Assembly.LoadFrom(binFile.Snapshots.First().SourceFile.AbsoluteFileName);
 
                 foreach (var type in assembly.GetExportedTypes())
                 {